# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transcription Service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ops/loadtest": {
      "post": {
        "summary": "Loadtest",
        "description": "Burst-submit synthetic jobs so the System panel shows the semaphore\nworking the vendor cap live. Same insert-then-enqueue path as /transcribe;\nthe high-water mark is reset so the burst is measured from zero.",
        "operationId": "loadtest_ops_loadtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoadTestRequest": {
        "properties": {
          "jobs": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Jobs",
            "default": 40
          },
          "chunks": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Chunks",
            "default": 8
          }
        },
        "type": "object",
        "title": "LoadTestRequest"
      },
      "LoadTestResponse": {
        "properties": {
          "jobsSubmitted": {
            "type": "integer",
            "title": "Jobssubmitted"
          },
          "chunksSubmitted": {
            "type": "integer",
            "title": "Chunkssubmitted"
          }
        },
        "type": "object",
        "required": [
          "jobsSubmitted",
          "chunksSubmitted"
        ],
        "title": "LoadTestResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}